+++ R/sym_symbol.R
@@ -79,6 +79,8 @@ eval_to_symbol <- function(x) {
   }
   
   x <- r_strings_to_python(x)
+  
+  # x <- "a/sqrt(a*conjugate(a))" fail
   s <- reticulate::py_eval(x, convert = FALSE)
   y <- construct_symbol_from_pyobj(s)
   return(y)