+++ NEWS.md
@@ -2,7 +2,7 @@
 
 * `print(..., method = "compactascii")`: column vector as transposed row vector
    and handling `caracas_scaled_matrix`
-* `LUdecomposition()` added.
+* `LUdecomposition()`, `chol()` (Cholesky), `svd_()` (Singular Value decomposition) added.
 * `cumsum()` added.
 
 # caracas 2.1.0